feature: split server.ts
